restructured assets into directories and updated the imports in corresponding files
