+++ client/src/assets/avatars/avatar2.png
