+++ client/src/assets/home/musicNotes.png
