+++ client/src/assets/home/studioHubLogo.png
