+++ client/src/assets/home/troubleshooting.png
